Allow not showing progress when sending an email.
When sharing we don't need to show the progress dialog because we
already have one and two of them look strange (also the message is
about sending email which might be confusing).

Ideally we should not use MailEditor for such purposes and should
extract all of this code into MailFacade or elsewhere.