[dev] Only redirect navigation requests to the start page.
We have a mechanism to load index.html when another "virtual" app path
is requested but we did not make distinction between requests so one
failing request could cause an infinite loop until the browser would
give up with too many redirects. This changes it so that it only
redirects requests for HTML not not JS requests. Maybe we should also
check that the file is there before doing it.