Fix tests to allow the owner of a group full permissions on it
The change in the code was introduced in
8409b80ad1aef4df7f344f794426982097c970e2 but the tests did not reflect
this.

The purpose of the function is to allow one to easily check whether a
user is allowed a certain operation on a group, not whether that
permission is explicitly written on the group. Therefore it was adapted
so that the owner of the group is allowed any operation that is checked.