Fix duplicated shortcuts, fix #1700
Shortcuts were duplicated because onbeforeremove mithril hook is only
called when element loses its parent. If the parent is removed, hook
is not called for children. This made us miss some of the
keyManager.unregisterShortcuts() calls and also some of the
removeFlash() calls.

onremove() on the other hand is called each time element is removed
from the DOM (whether via parent or directly) and that's what we need
in most cases.