Fix losing focus after opening contact form editor, fix #1977
Progress dialog tried to restore the focus to the element which had it
before the dialog but focus was already grabbed by the new dialog.

This moves opening of the contact form after the progress dialog so that
they are not open at the same time. Multiple modal dialogs do not play
well together and we should avoid it as much as possible.