Fix timing issue with facade proxy, fix #3753
WorkerProxy's exposeLocal was accepting raw queue for posting messages.
Queue and proxy are initialized before initial handshake is done so
that it is easier to manage dependencies. Our old
WorkerClient#_postRequest was already automatically waiting for
initialization but WorkerProxy was not. This lead to some
timing-dependent issues, e.g. deviceEncryptionFacade being requested
before any initialization is finished. Login would then get stuck.

This commit changes interface for exposeLocal so that it accepts a
dispatcher. On client side we pass a dispatcher which waits for init
process. Not only it is more flexible but also allows us to have
fewer types and get rid of downcasts.