[desktop] Fix accessing keychain and cancellation for it, fix #3733
There were few issues:

1. We would try to resolve any given key only once. If it fails, it
fails until we restart the app. Now we only cache successful
resolutions3

2. We would cache any of the keys forever, including credentials key.
We actually don't want to cache credential keys.

3. Using defer() led to a situation where we would treat error an as a
non-existing key and try to overwrite it. This meant that during
cancellation we would try to rewrite the key (which is bad by itself
and is exactly what we tried to avoid) and would also trigger another
prompt to unlock keychain.

4. Cancellation would not be propagated correctly and would lead to an
error on the client.

5. macOS would spam user with prompts for password event after other
changes. User has no idea what they give access to.

Now we only cache device key, we don't cache errors, and we don't
re-wrap errors (so that cancellation is propagated). We introduces
second service name for credentials so that user can differentiate
keys better (that's what macOS shows in prompt and in keychain list
instead of account name).