Don't call SendMailModel.saveDraft in parallel, fix #3726
It was possible to call saveDraft before the previous call has completed, by minimizing and re-opening the mail editor. This would cause for multiple drafts to be made of the same mail.

This commit forces all calls to saveDraft to wait for the previous call to complete before continuing with the next one. It also discards any intermediate calls, so only the most recent call will be queued up.

If a save call fails, then the promise gets reset but any followup calls will not be made

We also no longer pass in a wait callback for saveDraft. It is still necessary for send due to some underlying issue with the modal dialogs.