Bail out and delete credentials when we can't decrypt them, #3875
After running into race condition with keytar as part of #3875 we saved
some credentials which we can't decrypt anymore. To be precise, we
can't decrypt the key which we use for credentials but we can only
detect it when decrypting credentials themselves because only then we
have mac which doesn't match.

When this happens we treat crypto error as broken key and we invalidate
credentials.