Fix OutOfSync for persistent cache
The time when the entityEvents were downloaded last time was used
incorrectly. Because of how it was written and read it would always be
the same number, and we would never detect OutOfSync.

Now we moved it completely to the cache, and we rely on cache to keep
track of it. We do not write incorrect value before fetching the
events during initialization anymore.