Fix trying to load list while offline, fix #4032
With the new loading state for offline error we have been missing a
check in `loadMoreIfNecessary()` so the list loading would be triggering
continuously while scrolling near the retry button. It was especially
prominent on iOS where there is an overscroll effect and UI updates
seem to take longer.

Simply adding a check to `loadMoreIfNecessary()` is not enough as it
was also used for retry itself. Instead, we introduce another function
which doesn't check for position and only works in the correct state.