[android] improve slow attachment uploads by removing read timeout
due to TCP buffering we don't actually get to know when our data was
written to the network.
this may cause us to copy a file to the TCP buffer and immediately start
waiting on a server response.
on slow connections, this trips the 15s HTTP read timeout.

removing the timeout and relying on the OS to tell us if the connection
breaks should help with this.
for this, we convert all SocketExceptions into ConnectionErrors.
There's still sort of a timeout because the server will stop listening
after 10 (?) minutes to prevent slow loris.

also add a redraw on the minimized mail editor so we see a change in
status immediately

#4035