refactor file controller
FileController was a big bundle of logic and it was rather hard to understand what it would do at any given point.

This commit extracts 2 separate implementations, one for browser and one for native. Shared parts of the implementation were separated out into global functions

another thing to do would be to pull of the GUI calls out of the FileController, and also to simplify the implementation of FileControllerNative.downloadAll, it can probably just use doDownload