[desktop] fix ProtocolProxy not handling blob uploads
We have a case for closing ephemeral sessions via beacon request and
there we have to use Blob. This seems to be the only place where we
do upload request with blob. It was not handled before (we assumed that
we have a buffer).

Now we handle both blob and bytes.

close #4398