fix mailbody issues
fix the flicker when the unread state is changed:
Instead of recreating the view model we just update the mail in the existing one. This introduced the issue that the existing mail body would not be updated so we have to manually call `loadAll`, and also re-call `replaceInlineImages` hence the funky business with streams

fix the flickering when `contrastFixNeeded` is true: this is also fixed by not recreating the viewModel

fix the rendering of a new mailbody:
Mithril was reusing the dom element when switching from rendering a mail body to rendering the progress spinner. Because of the shadow dom, the old content was remaining. I opted to add keys to the different elements to force mithril to re-render it completely.

fix #4309