Have invoice page number not display on first page of signup
Specify an enabled function for InvoiceAndPaymentDataPage's attrs so
that its page button is displayed when appropriate.

When signing up for a new account, do not display it if the the user
either did not select Free or they are on the first page (and thus we
may not be certain if they want to select a paid plan).

We want to check this in particular for signup, as upgrading from free
to paid will, of course, always have an invoice page. And changing from
Paid to another plan does not have an invoice page to begin with.

Fixes #3100