fix unread mail counters not being corrected on full list load
previously, list would have a callback argument that got called when it
was fully loaded, which is where fixCounters was previously called.
With the removal of that argument in d7a14a, it was moved to the fetch
argument and called asynchronously in parallel.
It contains some checks to make sure we're setting the right value on
the counter, one of which caused us to almost always bail out after the
sequencing change:

* fetch
* get new items
* check if it's completely loaded and dispatch fixCounter if it is
* immediately return new items to be inserted in the list.
* run fixCounter, which would notice that the new items were inserted
  after dispatch and bail out because the list changed

With this commit, we add a marker of the last element
push/removal/replacement (better than checking the length).
It would be better to create some MonitoredArray class that can track
modifications instead of tracking the array and the marker as
separate fields, but this is the easiest and certainly more performant
than proxying array accesses.

With this commit, the mailListView will also re-dispatch the fixCounter
call if the list changed so we don't miss the chance to fix the counter.
There's a potential issue where the "fully loaded" state reverts between
the first and second call, but I couldn't figure out when and if this
might happen.

fix #4417

Co-authored-by: ivk <ivk@tutao.de>