Implement a button to toggle visibility for external passwords
The textfields for passwords when

- writing a mail to an external recipient
- adding a guest to a calendar event

are by default in plain text and have no button to toggle the
visibility. This behavior is caused by the TextFieldType
"ExternalPassword", however, in no instance should a password be shown
in plain text without the possibility to toggle this behavior.

The visibility of the aforementioned textfields can now be toggled with
a new button which by default assumes that the password is not visible.

Any usage of "TextType.ExternalPassword" was removed since it only
existed to work around the problem of browsers autofilling password
fields when not desired.

Co-authored-by: @nif-tutao
fixes #578