[desktop] Upload blobs via native
For reasons described in 6f4600f we started uploading
DataFile's via native. It was also changed in desktop but
we never actually implemented the desktop part.

With this change we fix multiple issues to make uploading blobs
work again on desktop:
- change writeDataFile() to not ask for path but instead write to a
temp directory for unencrypted files, like on other platforms,
and adjust the code accordingly
- implement aesEncryptFile(), splitFile() and upload(), needed for
blobs

fix #4529