Fix email body disappearing in specific case
Email body would disappear because the shadow dom container is
re-created on updated mail body but we would no-op if the mail body
didn't change. It seems like the container is re-created because we
delay mail body rendering when triggering mail update and it will
make mail body disappear for a short period of time.'

Now we make sure that on creation of the container we always render
mail body.

#4565