Remove usage test TTL behavior
This is because the current impl. was
causing bugs in PostLoginActions.onPartialLoginSuccess().
This method is also called on registering an account so
that the storage behavior was set to persist (although
the user is not quite logged in) and thus the ephemeral
tests were no longer accessible for the signup dropouts
or payment usage tests.