fix dates from before 1970 being truncated during offline serialization
we use CBOR blobs to encode our entities for storage in the offline DB.
This requires to tag items with type information, which was done
incorrectly for dates: we always set the type to uint, which is wrong
when storing the result of Date.getTime (ms since Jan 1 1970) for dates
before 1970 - these are negative. CBORG truncated these to 0, which
results in wrong start and end times after deserialization and failure
to display the events.

fix #4660