[ios] implement handling of share requests
the current implementation loads the item under the first supported
item type from each of the shared NSItemProviders. During our tests,
it was okay to ignore all but one item type because they seemed to be
different names for the same item.

after the items are loaded as NSSecureCoding, we try to convert them to
swift types that we can handle normally, filtering out any items where
that fails.

each item type has a handler that knows how to make it available to the
main app, which means either converting it to text to be added to the
mail body or writing it as a file in the app group folder to be
attached.

once everything is converted and writtten to disk, we create a
SharingInfo struct and serialize it to the shared UserDefaults.
It is saved with a timestamp as a key so we can distinguish several
concurrent share requests.
The main app is called by invoking its tutashare:// deep link handler
with the timestamp as the host portion of the link.

The main app gets the info and calls the createMailEditor ipc function.

Cleanup happens on app shutdown (deleting all shared content) and when
the creation of the mail editor fails for any reason (only data
associated with that particular request).

close #556

Co-authored-by: nok <nok@tutao.de>