add validation for U2F display name
since the spec says that hardware authenticators are allowed to
truncate display names and names to 64 bytes, we limit the length
of webauthn second factors to less than that.

on some platforms, this fixes a bug where we were unable to register
a webauthn second factor if the mail address or display name is too
long.

close #4727