fix flaky SendMailModel test
the mail model initialization can get a list of recipients that will be
resolved asynchronously by a method on the model.

this method may resolve after the rest of the initialization is done
and has cleared the saved state, which is unintended if the mail
was not actually changed between initialization and recipient list
resolution.

Co-authored-by: ivk <ivk@tutao.de>