Fix navigation in Mail view, fix view caching issue
After changing MailView to be a proper component we've introduced a
cache to hold on to some state. Unfortunately we had to hold the whole
MailListView in the cache because of how it is written. This was not
done properly because we create it only once and would hold on to the
first instance of MailView (which was used at creation time). If we
switch to another view and back we need to get a new MailView instance.

We fixed it by updating the MailView reference together with lifecycle.

A proper solution would mean not requiring MailView in there and not
holding on to the whole MailListView in the cache.'

fix #5026