External mailbox fixes
One issue with external mailbox was that the header was not given the
new attributes it needs. We've converted ExternalLoginView to a proper
component to help with that.

Another issue was assuming that all system folders exist for all
mailboxes. We've introduced another helper to reduce the boilerplate,
but we should be careful about using it.

fix #5031