Make ArchiveDataType not required for reading blobs via EntityRestClient
Requesting a read token no longer needs archiveDataType for archives
you own, so do not do this for MailDetailsBlob since each recipient
owns their own copy.

Also, if we add more blob element types, we should not assume that it
will be a MailDetails blob. This value will be ignored anyway in these
cases.

tutadb#1294