Add move shortcut to the SearchView
Before few places in the code used different and slightly invalid
checks for moving mails, e.g. single and multiple mail selections were
different, or the check for the same mailbox was done but the check for
the correct folder was omitted.

This change unifies the implementations to always use the same logic.

fix #4510