some improvements to credit card form
ranges are now used to identify a credit card type. Furthermore, we are
less strict on validating the credit card number. Hence, we allow
numbers which are longer than the recognized credit card type allows
since the credit card type can't be derived with 100% accuracy.

cvv is now called the same in all places depending on the derived credit
card type

#5055