Fix notifying about opened email twice
onEmailOpened() was called twice: once in createConversationViewModel()
and once right next to it in onListStateChange(). Because it was
happening so quickly some buffering would take place and both
invocations would be sent as one message.

now onEmailOpened() is not called in createConversationViewModel()
anymore but only outside of it so there are no double invocations.

fix #5545