Fix subscription plan boxes not fitting into the dialog
The issue appeared because of the scrollbar changes. The effective
space in the dialog got reduced and the boxes wouldn't fit anymore.

We noticed that there's plenty of space around the subscription boxes
and that there's also enough padding on the dialog edges so we just
made sure there's a gap between boxes.

fix #5598