fix sanitizeHtml removing tutatemplate: links
the regex allowing the tutatemplate: protocol
was only added to the sanitizeFragment config.

using sanitizeFragment for sanitizing knowledge
base entries would require us to insert the result
manually into the DOM since mithril does not support
trusting fragments and we don't want to serialize
to a string and then trust.

so we add the regex to the sanitizeHtml config as well

fix #5640