fix altered instances not being excluded on import
we're now grouping events by uid before checking if they
should be created or not.
the first event of a group without a recurrenceId
is the designated progenitor (the rest is dropped
as duplicate). if it has a repeat rule and we
encounter altered instances in that group,
their recurrenceIds are added as excluded instances
on the progenitor.

this does not handle updating existing progenitors
when new altered instances are imported or adding
existing altered instances on newly imported
progenitors

fix #5708
fix ed27db58b2735268d70a6119ed1deeb27e8a8919